BobArnso: Expose DTF's Database.ExecutePropertyQuery method.
RobMen: Support retrying package install in Burn.
        Add balretry to convert common package errors to retry recommendations.

--HG--
branch : wix36

+++ src/dutil/thmutil.cpp
@@ -2421,7 +2421,7 @@ static HRESULT ParseControl(
         hr = XmlGetAttributeNumber(pixn, L"Interval", &dwValue);
         if (S_OK == hr && dwValue)
         {
-            pControl->wBillboardInterval = dwValue & 0xFFFF;
+            pControl->wBillboardInterval = static_cast<WORD>(dwValue & 0xFFFF);
         }
         ExitOnFailure(hr, "Failed to get Billboard/@Interval.");
